Moved canceling of job to KubernetesLauncher
